backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
SQLAlchemy
psycopg2-binary
passlib==1.7.4
python-multipart
python-jose[cryptography]
email-validator
google-generativeai
httpx>=0.27

